requirements.txt
astropy==7.2.0
astropy-iers-data==0.2026.2.16.0.48.25
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dash==4.0.0
Flask==3.1.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.16.0
nest_asyncio==1.6.0
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
photutils==2.3.0
plotly==6.5.2
pyerfa==2.0.1.5
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
scipy==1.17.0
six==1.17.0
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
